Move Publisher class to Bind
Remove spec examples that were mocking argument passing (they weren't testing much, anyway)
